+++ instrumentation/micrometer/micrometer-1.5/library/src/main/java/io/opentelemetry/instrumentation/micrometer/v1_5/OpenTelemetryMeterRegistryBuilder.java
@@ -17,10 +17,16 @@ import io.micrometer.core.instrument.config.NamingConvention;
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.metrics.MeterBuilder;
 import io.opentelemetry.instrumentation.api.internal.EmbeddedInstrumentationProperties;
+import io.opentelemetry.instrumentation.micrometer.v1_5.internal.Experimental;
 import java.util.concurrent.TimeUnit;
 
 /** A builder of {@link OpenTelemetryMeterRegistry}. */
 public final class OpenTelemetryMeterRegistryBuilder {
+
+  static {
+    Experimental.internalSetMicrometerHistogramGaugesEnabled(
+        (builder, enabled) -> builder.histogramGaugesEnabled = enabled);
+  }
 
   // Visible for testing
   private static final String INSTRUMENTATION_NAME = "io.opentelemetry.micrometer-1.5";
@@ -75,7 +81,12 @@ public final class OpenTelemetryMeterRegistryBuilder {
    *
    * <p>This is disabled by default, set this to {@code true} to enable gauge-based Micrometer
    * histograms.
+   *
+   * @deprecated Use {@link
+   *     io.opentelemetry.instrumentation.micrometer.v1_5.internal.Experimental#setMicrometerHistogramGaugesEnabled(OpenTelemetryMeterRegistryBuilder,
+   *     boolean)} instead. This method will be removed in the next release.
    */
+  @Deprecated // will be removed in the next release
   @CanIgnoreReturnValue
   public OpenTelemetryMeterRegistryBuilder setMicrometerHistogramGaugesEnabled(
       boolean histogramGaugesEnabled) {